More UFS spring cleaning - move fs.h into same fodler as other ufs bits
