+++ sys/src/libufs/cgroup.c
@@ -30,7 +30,7 @@
 
 #include <ufs/ufsdat.h>
 #include <ufs/dinode.h>
-#include <ffs/fs.h>
+#include <ufs/fs.h>
 #include <ufs/libufs.h>
 
 ufs2_daddr_t
